fix: ignore non-functional smartcard readers
This includes another non-obvious fix:

I noticed that we connect *twice* to the reader because
`key::filter_connected` already opens a connection - which we then don't
use. First of all, it's not necessary to check for connected readers
because the subsequent `open_connection` will fail anyway and we just
skip it. **More importantly, it seems that the connection created using
filter_connected invalidates the pin session** because it's not closed
using `disconnect_without_reset`.